cleaning up test file
